extra modules cleanup
