+++ customics/__init__.py
@@ -1,6 +1,6 @@
 import importlib.metadata
 
-from . import datasets, metrics, tools
+from . import datasets, metrics
 from .model import CustOMICS
 
 __version__ = importlib.metadata.version("customics")
