+++ customics/modules/__init__.py
@@ -1,2 +1,5 @@
+from .fully_connected import FullyConnectedLayer
+from .encoders import Encoder, ProbabilisticEncoder
+from .decoders import Decoder, ProbabilisticDecoder
 from .autoencoder import AutoEncoder
 from .vae import VAE
