--- customics/tools/__init__.py
@@ -1,2 +1,0 @@
-from .utils import get_common_samples, get_sub_omics_df
-from .net_utils import FullyConnectedLayer
